requirements.txt
apify-client
requests
google-genai
python-dotenv
pytz
yfinance
pandas
